requirements.txt
python-dotenv==1.0.0
SQLAlchemy==1.4.41
Flask==2.1.1
